Fixed installation to only include cairo files

+++ felucca/utils.py
@@ -51,15 +51,17 @@ def get_package_name() -> str:
 
 
 def install_cairo_package(package):
-    def get_ignored_files(d, filenames):
-        return filter(lambda fname: not fname.endswith("cairo"), filenames)
 
     site_packages = site.getsitepackages()[0]
     norm_package = package.replace("-", "_")
     source_package = get_package_name().replace("-", "_")
     package_contracts = os.path.join(site_packages, norm_package)
     target_dir = f"./{source_package}/{norm_package}"
-    shutil.copytree(package_contracts, target_dir, ignore=get_ignored_files)
+    shutil.copytree(
+        package_contracts,
+        target_dir,
+        ignore=shutil.ignore_patterns("__pycache__", "*.py"),
+    )
 
     for contract in Path(target_dir).glob("**/*.cairo"):
         content = contract.read_text()
@@ -73,4 +75,5 @@ def uninstall_cairo_package(package):
 
     norm_package = package.replace("-", "_")
     target_dir = f"./{get_package_name()}/{norm_package}"
-    shutil.rmtree(target_dir)
+    if os.path.exists(target_dir) and os.path.isdir(target_dir):
+        shutil.rmtree(target_dir)
